feat(#302): Refactor terms into individual classes for cleaner design
