+++ lib/factbase/terms/concat.rb
@@ -4,11 +4,12 @@
 # SPDX-License-Identifier: MIT
 
 require_relative 'base'
-# The Factbase::Unique class provides functionality for evaluating the uniqueness
-# of terms based on provided operands and facts.
-class Factbase::Prev < Factbase::TermBase
-  # Constructor.
-  # @param [Array] operands Operands
+# Concat term.
+#
+# Author:: Yegor Bugayenko (yegor256@gmail.com)
+# Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
+# License:: MIT
+class Factbase::Concat < Factbase::TermBase
   def initialize(operands)
     super()
     @operands = operands
@@ -18,12 +19,8 @@ class Factbase::Prev < Factbase::TermBase
   # @param [Factbase::Fact] fact The fact
   # @param [Array<Factbase::Fact>] maps All maps available
   # @param [Factbase] fb Factbase to use for sub-queries
-  # @return [Object] The previous value
+  # @return [String] The concatenated string
   def evaluate(fact, maps, fb)
-    assert_args(1)
-    before = @prev
-    v = _values(0, fact, maps, fb)
-    @prev = v
-    before
+    (0..(@operands.length - 1)).map { |i| _values(i, fact, maps, fb)&.first }.join
   end
 end
