+++ test/factbase/terms/test_matches.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+
+# SPDX-FileCopyrightText: Copyright (c) 2024-2025 Yegor Bugayenko
+# SPDX-License-Identifier: MIT
+
+# Test for unique term.
+# Author:: Volodya Lombrozo (volodya.lombrozo@gmail.com)
+# Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
+# License:: MIT
+
+require_relative '../../test__helper'
+require_relative '../../../lib/factbase/term'
+require_relative '../../../lib/factbase/terms/matches'
+
+class TestMatches < Factbase::Test
+  def test_regexp_matching
+    t = Factbase::Matches.new([:foo, '[a-z]+'])
+    assert(t.evaluate(fact('foo' => 'hello'), [], Factbase.new))
+    assert(t.evaluate(fact('foo' => 'hello 42'), [], Factbase.new))
+    refute(t.evaluate(fact('foo' => 42), [], Factbase.new))
+  end
+end
